fix(contamination): 除染が進まない/新ゾーンが即消滅する不具合を修正
Codex レビュー指摘（P1/P2）の修正。

- P1: 濃度を byte で持ち微小間隔ごとに丸めていたため、5%/月の相対除去が丸めで停滞し
  ゾーンが減衰・消滅しなかった。濃度を float 化し減衰係数を掛け続けることで端数を失わず着実に減衰させる
  （ContaminationZone.Intensity=float、ContaminationDecay.DecayFactor、ZoneSerializer v3）
- P2: ゾーンが無い間に経過時計を進めず、後発ゾーンに空白期間を丸ごと課して即除染消滅していた。
  ゾーン有無に関わらず処理サイクルで時刻を前進させるよう修正

Coreテスト135件全緑。

+++ src/MissileDisaster/Core/ContaminationZone.cs
@@ -1,9 +1,9 @@
 namespace MissileDisaster.Core
 {
     /// <summary>
-    /// 放射能汚染ゾーン。ワールド座標中心・半径(m)・発生ゲーム内時刻(Ticks)・現在の濃度(0-255)。
-    /// Intensity は除染施設により恒久的に低下する（下がった濃度で維持され、元には戻らない）。
-    /// NuclearMeltdown.Core.ContaminationZone を拡張移植。UnityEngine 非依存。
+    /// 放射能汚染ゾーン。ワールド座標中心・半径(m)・発生ゲーム内時刻(Ticks)・現在濃度(0-255)。
+    /// Intensity は float（除染で連続的に低下し、微小な減衰も端数として蓄積される）。土壌汚染フィールドへ
+    /// 書き込む際に整数へ丸める。除染で下がった濃度は元へ戻らない。UnityEngine 非依存。
     /// </summary>
     public struct ContaminationZone
     {
@@ -11,14 +11,14 @@ namespace MissileDisaster.Core
         public float CenterZ;
         public float Radius;
         public long StartTicks;
-        public byte Intensity;
+        public float Intensity;
 
         public ContaminationZone(float centerX, float centerZ, float radius, long startTicks)
-            : this(centerX, centerZ, radius, startTicks, 255)
+            : this(centerX, centerZ, radius, startTicks, 255f)
         {
         }
 
-        public ContaminationZone(float centerX, float centerZ, float radius, long startTicks, byte intensity)
+        public ContaminationZone(float centerX, float centerZ, float radius, long startTicks, float intensity)
         {
             CenterX = centerX;
             CenterZ = centerZ;
